+++ tests/testthat/test-r2_mcfadden.R
@@ -16,5 +16,11 @@ if (require("testthat") && require("performance") && require("MASS")) {
         model_type = "Generalized Linear", class = "r2_generic"
       ),
       tolerance = 1e-3)
+
+    expect_equal(
+      r2(model),
+      list(R2_Nagelkerke = c(`Nagelkerke's R2` = 0.108408289848161)),
+      tolerance = 1e-3
+    )
   })
 }