+++ NEWS.md
@@ -12,6 +12,8 @@
 * `performance_logloss()` to compute the log-loss of models with binary outcome. The log-loss is a proper scoring function comparable to the `rmse()`.
 * `performance_pcp()` to calculate the percentage of correct predictions for models with binary outcome.
 * `r2_mckelvey()` to compute McKelvey and Zavoinas R2 value.
+* `r2_zeroinflated()` as a crude R2 measure for zero-inflated models.
+* `r2_xu()` as a crude R2 measure for linear (mixed) models.
 
 ## Breaking changes